requirements.txt
pandas
pytrends
playwright
matplotlib
seaborn
statsmodels
python-dotenv
kagglehub